updated to just use a single lookup table from minifloat to single.
